% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataAnalysis.R
\name{dataAnalysis}
\alias{dataAnalysis}
\title{Analyze the Data}
\usage{
dataAnalysis(
  phenofinal = NULL,
  betafinal = NULL,
  array = "EPIC",
  Table1vars = c("Age", "BMI", "Smoke", "Parity", "Sex", "Gestage", "BWT"),
  StratifyTable1 = FALSE,
  StratifyTable1var = NULL,
  maxit = 100,
  vartype = "ExposureCont",
  varofinterest = NULL,
  adjustmentvariables = c("Age", "Gestage"),
  Omega = NULL,
  RestrictToSubset = FALSE,
  RestrictionVar = NULL,
  RestrictToIndicator = NULL,
  robust = TRUE,
  RunUnadjusted = TRUE,
  RunAdjusted = TRUE,
  RunCellTypeAdjusted = TRUE,
  RunSexSpecific = TRUE,
  RunCellTypeInteract = TRUE,
  runparallel = FALSE,
  number_cores = NULL,
  destinationfolder = NULL,
  savelog = TRUE,
  cohort = NULL,
  analysisdate = NULL,
  analysisname = NULL
)
}
\arguments{
\item{phenofinal}{A data.frame of phenotype data that is subset to the samples
among the matrix of processed beta-values. Be sure that the dataframe
phenofinal includes the column 'Sex', which can only take three possible
character values: 'Female', 'Male', or 'Unknown'. Missing values are not
allowed. The data.frame must also have the column 'Basename'. If
RestrictToSubset is TRUE, this data.frame must also include the column
specified in the argument RestrictionVar}

\item{betafinal}{Matrix of the processed beta-values}

\item{array}{A character string of array type; options: "EPIC" or "450K"}

\item{Table1vars}{An character vector of variable column names to include in
the descriptive Table 1. The column names must correspond the column names
in the pData DataFrame of the GenomicRatioSet (e.g. c("Age", "BMI",
"Smoke")); see the function loadingSamples for the new column names}

\item{StratifyTable1}{Logic statement (TRUE or FALSE), whether to stratify the
Table 1 data on a specific variable like infant sex or case/control status;
default is FALSE}

\item{StratifyTable1var}{A character string indicating the column name for the
variable on which to stratify the Table 1 data; default is FALSE. Must be
specified if StratifyTable1 is TRUE}

\item{maxit}{If running robust linear regression models, the maximum number of
iterations (default=100)}

\item{vartype}{A character string indicating how to model DNA methylation;
four options: "ExposureCont", "ExposureCat", "OutcomeCont", "OutcomeBin". If
"ExposureCont", the variable of interest (varofinterest) is modeled as a
continuous exposure and assumed to be numeric. If "ExposureCat", the
variable of interest is modeled as a categorical exposure and assumed to be
a factor. If "OutcomeCont", the variable of interest is modeled as the
outcome, and assumed to be continuous. If "OutcomeBin", the variable of
interest is modeled as the outcome, and assumed to be binary (values are 0
or 1). If DNAm is modeled as the exposure, the site-specific effect
estimates will be for a 0.01 increase in beta-value}

\item{varofinterest}{A character string of the column name of the exposure or
outcome of interest (e.g. "BWT"). The column name must correspond to a
column name in the pData DataFrame of the GenomicRatioSet; see the function
loadingSamples for the new column names}

\item{adjustmentvariables}{A character vector of variable column names to
include in your multivariable adjusted models (e.g. c("Age","Gestage")). The
column names must correspond the column names in the pData DataFrame of the
GenomicRatioSet; see the function loadingSamples for the new column names}

\item{Omega}{A required matrix of cell composition estimates generated by the
preprocessingofData function}

\item{RestrictToSubset}{Logic statement (TRUE or FALSE), whether to restrict
the analysis to a specific subset of the samples. The default is FALSE, in
other words, do not restrict the analysis to a specific subset of the
samples. If specified TRUE, must provide a character string to the arguments
RestrictionVar and RestrictToIndicator; see those arguments for additional
details.}

\item{RestrictionVar}{A character string indicating the column name for the
variable in the phenofinal data.frame that will be used to subset the data.
Must be specified if RestrictToSubset is TRUE; default is NULL}

\item{RestrictToIndicator}{A character string indicating the category of the
RestrictionVar you would like to restrict the analysis to. This character
string must equal one of the RestrictionVar categories. This argument is
only required if RestrictToSubset = TRUE; default is NULL}

\item{robust}{Logic statement (TRUE or FALSE), whether to run robust linear
regressions. The default is TRUE}

\item{RunUnadjusted}{Run unadjusted models? Logic statement (TRUE or FALSE),
default = TRUE}

\item{RunAdjusted}{Run adjusted models? Logic statement (TRUE or FALSE),
default = TRUE}

\item{RunCellTypeAdjusted}{Run adjusted analysis, including cell composition?
Logic statement (TRUE or FALSE), default = TRUE}

\item{RunSexSpecific}{Run models stratifying by infant sex? Logic statement
(TRUE or FALSE), default = TRUE}

\item{RunCellTypeInteract}{Whether to run CellDMC to identify the specific
cell type(s) driving the differential methylation. Logic statement (TRUE or
FALSE), default = TRUE}

\item{runparallel}{Whether to run the site-specific analysis in parallel (TRUE
or FALSE), default = FALSE}

\item{number_cores}{If running in parallel, a numeric string indicating the
number of cores to run (e.g. 4), default = NULL}

\item{destinationfolder}{A character string indicating the location where
files should be saved, e.g. "C:\\Home\\PACE\\BirthSize"}

\item{savelog}{logic; TRUE indicates to save a log of the functions run
(default)}

\item{cohort}{A character string for the cohort's acronym (e.g. "HEBC")}

\item{analysisdate}{A character string indicating the date the analysis was
run. Please specify in the form: YEARMONTHDAY, e.g. "20200205" for February
5th 2020}

\item{analysisname}{An optional character string for the analysis name. This
will be used to label the output directory and the RData files that contain
the final results. This can be helpful to specify if running additional
sensitivity analyses}
}
\value{
Outputs a list of data frames with the site-specific association
 results. Also creates a new folder in the Output folder for the specified
 exposure/outcome variable. The folder includes: \enumerate{\item Descriptive
 statistics of cohort (Table 1; csv file) \item Figures (PDFs/PNGs) of
 associations, including: \itemize{\item QQplot for each analysis \item
 Volcano plot for each analysis \item Scatterplot for the 20 most significant
 associations for each analysis} \item An RData file that includes a list of
 data frames with the site-specific association results \item An RData file
 that includes a list of the lambdas for each analysis}
}
\description{
The goal is to output the information necessary for meta-analysis
}
\details{
Data analysis steps include: \enumerate{ \item Create a Table 1
 descriptive table for the cohort \item Test the association between
 estimated cellular components and the exposure variable, the adjustment
 variables, and batch (if included)  \item Perform association analysis
 between variable of interest and each CpG site individually. If vartype is
 "ExposureCont", "ExposureCat", or "OutcomeCont", the association is
 estimated using linear regression modelling (if robust=TRUE, using the rlm()
 function in R; otherwise, using lm()). If vartype is "OutcomeBin", a
 logistic regression is used to estimate the association
 (glm(family="binomial")). Analyses performed include: \itemize{ \item
 Unadjusted analysis if RunUnadjusted=TRUE \item Analysis adjusting for
 specified adjustment variables if RunAdjusted=TRUE \item Analysis adjusting
 for specified adjustment variables and estimated cell composition if
 RunCellTypeAdjusted=TRUE \item If RunSexSpecific=TRUE, RunUnadjusted=TRUE,
 and there are at least 10 female infants in the dataset, unadjusted analysis
 among the female infants \item If RunSexSpecific=TRUE, RunAdjusted=TRUE, and
 there are at least 10 female infants in the dataset, analysis adjusting for
 specified adjustment variables among the female infants \item If
 RunSexSpecific=TRUE, RunCellTypeAdjusted=TRUE, and there are at least 10
 female infants in the dataset, analysis adjusting for specified adjustment
 variables and estimated cell composition among the female infants \item If
 RunSexSpecific=TRUE, RunUnadjusted=TRUE, and there are at least 10 male
 infants in the dataset, unadjusted analysis among the male infants \item If
 RunSexSpecific=TRUE, RunAdjusted=TRUE, and there are at least 10 male
 infants in the dataset, analysis adjusting for specified adjustment
 variables among the male infants \item If RunSexSpecific=TRUE,
 RunCellTypeAdjusted=TRUE, and there are more than 10 male infants in the
 dataset, analysis adjusting for specified adjustment variables and estimated
 cell composition among the male infants \item If RunCellTypeInteract=TRUE,
 output of CellDMC (from package EpiDISH) based on the adjusted model}  }
}
\examples{
\dontrun{
allresults<-dataAnalysis(phenofinal=pData(processedOut$mset),
                        betafinal=processedOut$processedBetas,
                        array="EPIC",
                        Omega=processedOut$Omega,
                        StratifyTable1=TRUE,
                        StratifyTable1var="Sex",
                        maxit=100,
                        Table1vars=c("Age","BMI","Smoke","Parity","Sex","Gestage","BWT"),
                        vartype="ExposureCont",
                        varofinterest="BWT",
                        adjustmentvariables=c("Age","Gestage"),
                        destinationfolder="H:/UCLA/PACE/Birthweight-placenta",
                        robust=TRUE,
                        RunUnadjusted=TRUE,
                        RunAdjusted=TRUE,
                        RunCellTypeAdjusted=TRUE,
                        RunSexSpecific=TRUE,
                        RunCellTypeInteract=TRUE,
                        RestrictToSubset=FALSE,
                        RestrictionVar=NULL,
                        RestrictToIndicator=NULL,
                        runparallel=FALSE,
                        number_cores=NULL,
                        savelog=TRUE,
                        cohort="HEBC",analysisdate="20200710",
                        analysisname="primaryBWT")
}
}
