% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingSamples.R
\name{loadingSamples}
\alias{loadingSamples}
\title{Load Data and Rename Phenotype Data to Standardize}
\usage{
loadingSamples(
  SamplePlacement = NULL,
  PhenoData = PhenoData,
  IDlink = "ID",
  BWTvar = NULL,
  BATCHvar = NULL,
  SEXvar = NULL,
  FemaleInd = "1",
  MaleInd = "2",
  ETHNICvar = NULL,
  GESTvar = NULL,
  BIRTHLENGTHvar = NULL,
  HEADCIRCUMvar = NULL,
  IDATdir = NULL,
  destinationfolder = NULL,
  savelog = TRUE,
  cohort = NULL,
  analysisdate = NULL
)
}
\arguments{
\item{SamplePlacement}{An data.frame that includes 'Basename' and sample ID.
Not required if 'Basename'is in PhenoData; might also include an indicator
for batch.}

\item{PhenoData}{A required data.frame that includes the participant
characteristics and a sample ID}

\item{IDlink}{A required character string indicating the column name for the
sample ID in the data.frame PhenoData}

\item{BWTvar}{An optional character string indicating the column name for the
birthweight in the data.frame PhenoData; if there is no birthweight variable
in data.frame PhenoData, can specify argument value is NULL. Birthweight
values should be in grams}

\item{BATCHvar}{An optional character string indicating the column name for
the batch variable in the data.frame PhenoData; if there is no batch
variable in the data.frame PhenoData, can specify argument value is NULL.
Batch variable is expected to be categorical}

\item{SEXvar}{A required character string indicating the column name for the
infant in the data.frame PhenoData; assumed only two sexes are in the
dataset}

\item{FemaleInd}{A character string indicating how female sex is coded in the
dataset; e.g. "1"}

\item{MaleInd}{A character string indicating how male sex is coded in the
dataset; e.g. "2"}

\item{ETHNICvar}{An optional character string indicating the column name for
the maternal race/ethnicity variable in the data.frame PhenoData; if there
is no maternal race/ethnicity variable in the data.frame PhenoData, can
specify argument value is NULL. Maternal race/ethnicity variable is expected
to be categorical}

\item{GESTvar}{An optional character string indicating the column name for the
gestational age variable in the data.frame PhenoData; if there is no
gestational age variable in the data.frame PhenoData, can specify argument
value is NULL. Gestational age variable is expected to be in days}

\item{BIRTHLENGTHvar}{An optional character string indicating the column name
for the birth length variable in the data.frame PhenoData; if there is no
birth length variable in the data.frame PhenoData, can specify argument
value is NULL. Birth length variable is expected to be in centimeters}

\item{HEADCIRCUMvar}{An optional character string indicating the column name
for the head circumference variable in the data.frame PhenoData; if there is
no head circumference variable in the data.frame PhenoData, can specify
argument value is NULL. Head circumference variable is expected to be in
centimeters}

\item{IDATdir}{A character string indicating the location of the IDAT files,
e.g. "C:/Home/PACE/BirthSize/IDATs"}

\item{destinationfolder}{A character string indicating the location where
files should be saved, e.g. "C:/Home/PACE/BirthSize"}

\item{savelog}{logic; TRUE indicates to save a log of the functions run
(default)}

\item{cohort}{A character string for the cohort's acronym (e.g. "HEBC")}

\item{analysisdate}{A character string indicating the date the analysis was
run. Please specify in the form: YEARMONTHDAY, e.g. "20200205" for February
5th 2020}
}
\value{
A RGset (of class RGChannelSetExtended) with raw microarray intensity
 values. In addition to the columns included in the input PhenoData
 data.frame, the pData for this RGset will include the following columns:
 \item{ID}{Sample ID} \item{BWT}{Birthweight variable if BWTvar argument was
 not NULL; numeric} \item{Age}{Maternal age variable if AGEvar argument was
 not NULL; numeric} \item{BMI}{Maternal pre-pregnancy variable if BMIvar
 argument was not NULL; numeric} \item{Sex}{Infant sex variable if SEXvar
 argument was not NULL; factor, 'Female' or 'Male'} \item{Ethnic}{Maternal
 race/ethnicity variable if ETHNICvar argument was not NULL; factor}
 \item{Gestage}{Gestational age variable if GESTvar argument was not NULL;
 numeric}\item{Batch}{Batch variable if BATCHvar argument was not NULL;
 factor} \item{BirthLength}{Birth length variable if BIRTHLENGTHvar argument
 was not NULL; numeric} \item{HeadCircum}{Head circumference variable if
 HEADCIRCUMvar argument was not NULL; numeric} \item{BWTkg}{Birthweight in kg
 if BWTvar argument was not NULL} \item{wlr}{Birthweight birth length ratio
 if BWTvar and BirthLength arguments were not NULL}
 \item{BWT_Zscore}{Birthweight gestational age and sex-specific Z-scores
 based on Newborn Cross-Sectional Study of the INTERGROWTH-21st Project}
 \item{BirthLength_Zscore}{Birth length gestational age and sex-specific
 Z-scores based on Newborn Cross-Sectional Study of the INTERGROWTH-21st
 Project} \item{HeadCircum_Zscore}{Head circumference gestational age and
 sex-specific Z-scores based on Newborn Cross-Sectional Study of the
 INTERGROWTH-21st Project} \item{WLR_Zscore}{Birthweight birth length ratio
 gestational age and sex-specific Z-scores based on Newborn Cross-Sectional
 Study of the INTERGROWTH-21st Project}
}
\description{
Reorganize data to standardize for analysis
}
\details{
This function Renames the variables, puts the variables in the right
 mode for the analysis, and loads the raw array data.
}
\examples{
\dontrun{
exampledat<-loadingSamples(SamplePlacement=NULL,PhenoData=allphenodata,IDlink="ID",
                          BWTvar="BWT",BATCHvar="Study",
                          SEXvar="GENDER_A",FemaleInd="1",MaleInd="2",
                          ETHNICvar="ETHNIC",GESTvar="gestAge",
                          BIRTHLENGTHvar="BirthLength",HEADCIRCUMvar=NULL,
                          IDATdir="H:/UCLA/PACE/Birthweight-placenta/IDATfiles",
                          destinationfolder="H:/UCLA/PACE/Birthweight-placenta",
                          savelog=TRUE,
                          cohort="HEBC",analysisdate="20200710")
}
}
