% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessingofData.R
\name{preprocessingofData}
\alias{preprocessingofData}
\title{Perform Sample Pre-processing}
\usage{
preprocessingofData(
  RGset = RGset,
  SamplestoRemove = NULL,
  ProbestoRemove = NULL,
  DetectionPvals = NULL,
  compositeCellType = "RefFree",
  KchooseManual = NULL,
  cohort = cohort,
  analysisdate = analysisdate,
  destinationfolder = NULL,
  savelog = TRUE
)
}
\arguments{
\item{RGset}{The RGset generated by the function loadingSamples. If you want
to adjust for batch effects using ComBat, the pData of the RGset must
include the column 'Batch'}

\item{SamplestoRemove}{An optional character vector of Basenames to exclude
based on the exploratory data analysis; default is NULL}

\item{ProbestoRemove}{An optional character vector of CpGIDs to exclude based
on the exploratory data analysis; default is NULL}

\item{DetectionPvals}{An optional matrix where DNA methylation intensities are
coded as NA (missing) if they do not meet the detection p-value, and 1
otherwise. This matrix is returned by the function ExploratoryDataAnalysis}

\item{compositeCellType}{A required character string indicating the type of
cell composition estimation to perform. The default is "RefFree" indicating
a reference-free deconvolution; other options are "Blood", "Placenta", and
"CordBlood" to be used with the Houseman reference-based approach. The
placenta reference-based approach is based on the data curated in the
following publication: https://www.researchsquare.com/article/rs-38223/v3}

\item{KchooseManual}{Optional numeric value (e.g 2) used to manually set the K
(number of cell types) estimated using RefFreeCellMixArray. May need to do
this if it seems that there is an extreme outlier driving the number of cell
types estimated. See details for more information; the default value is NULL}

\item{cohort}{A character string for the cohort's acronym (e.g. "HEBC")}

\item{analysisdate}{A character string indicating the date the analysis was
run. Please specify in the form: YEARMONTHDAY, e.g. "20200205" for February
5th 2020}

\item{destinationfolder}{A character string indicating the location where
files should be saved, e.g. "C:\\Home\\PACE\\BirthSize"}

\item{savelog}{logic; TRUE indicates to save a log of the functions run
(default)}
}
\value{
Figures (PDFs/PNGs) of associations between top PCs and indicators of
 batch before and after ComBat (if an indicator for batch was included in the
 dataset), as well as plots to identify potential outliers driving the
 estimated cell composition estimates. Also saves an RData file and returns a
 list that includes: \item{mset}{A GenomicRatioSet subset to the samples and
 CpG loci among the matrix of processed beta-values}
 \item{processedBetas}{Matrix of the processed beta-values} \item{Kchoose}{A
 numeric value indicating the number of cell types selected by the RefFree
 method}\item{Omega}{A matrix of cell composition estimates generated by the
 RefFree method}
}
\description{
Sample pre-processing so that the data is ready for analysis
}
\details{
Pre-processing steps include: \enumerate{ \item Drop samples and CpG
 loci based on exploratory data analysis \item Dye-bias correction by linear
 scaling based on normalization control probes on both color channels (minfi
 approach), followed by Noob (normal-exponential using out-of-band probes)
 background correction. \item Functional normalization \item Beta-mixture
 quantile (BMIQ) normalization method to correct for the bias of type-2 probe
 values \item Correction for batch effects (if indicated) using ComBat \item
 Estimate cellular heterogeneity with the RefFree method (Houseman et al) or
 reference-based approaches. After estimating RefFree Omega, examine
 OutlierScreening plots generated by this function to ensure no extreme
 outliers are being generated. If so, reduce Kchoose value by 1 using the
 argument KchooseManual, re-run the function preprocessingofData, and
 re-examine outlier plots. Reduce KchooseManual until extreme outliers are
 gone. \item Remove extreme outliers using a modified version of GapHunter.
 Gaps must be at least 3*IQR; the cutoff for the number of outliers in a
 group is either a maximum of 5 or 0.0025 of the total number of samples
 (whichever is larger). Detected outliers are recoded as NA.}
}
\examples{
\dontrun{
processedOut<-preprocessingofData(RGset=exampledat,
                                 SamplestoRemove=EDAtrying$SamplestoRemove,
                                 ProbestoRemove=EDAtrying$ProbestoRemove,
                                 DetectionPvals=EDAtrying$DetectionPvals,
                                 compositeCellType ="RefFree",
                                 KchooseManual=NULL,
                                 destinationfolder="H:/UCLA/PACE/Birthweight-placenta",
                                 savelog=TRUE,
                                 cohort="HEBC",analysisdate="20200710")
}
}
