% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sesamize.R
\name{RGChannelSetToSigDFs}
\alias{RGChannelSetToSigDFs}
\title{Convert RGChannelSet (minfi) to a list of SigDF (SeSAMe)}
\usage{
RGChannelSetToSigDFs(
  rgSet,
  manifest = NULL,
  BPPARAM = BiocParallel::SerialParam()
)
}
\description{
Convert RGChannelSet (minfi) to a list of SigDF (SeSAMe)
}
