% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExploratoryDataAnalysis.R
\name{ExploratoryDataAnalysis}
\alias{ExploratoryDataAnalysis}
\title{Exploratory Data Analysis}
\usage{
ExploratoryDataAnalysis(
  RGset = RGset,
  globalvarexplore = NULL,
  destinationfolder = NULL,
  savelog = TRUE,
  cohort = NULL,
  analysisdate = NULL
)
}
\arguments{
\item{RGset}{The RGset generated by the function loadingSamples. The pData
data.frame requires a column called "ID", which is a unique identifier for
each sample. The pData data.frame also requires a column called "Sex", which
can only take three possible character values: 'Female', 'Male', or
'Unknown'.}

\item{globalvarexplore}{A vector of characteristic column names to explore in
association with the first 12 principle components. These column names must
be in the pData frame of the RGset specified by the RGset argument}

\item{destinationfolder}{A character string indicating the location where
files should be saved, e.g. "C:\\Home\\PACE\\BirthSize"}

\item{savelog}{logic; TRUE indicates to save a log of the functions run
(default)}

\item{cohort}{A character string for the cohort's acronym (e.g. "HEBC")}

\item{analysisdate}{A character string indicating the date the analysis was
run. Please specify in the form: YEARMONTHDAY, e.g. "20200205" for February
5th 2020}
}
\value{
PDFs/PNGs of exploratory data analysis in "EDA" subfolder of "Output"
 folder, including: red/green signal intensities, distribution of intensity
 across all sites, heatmap of chrY, signal intensities on X and Y
 chromosomes, dendogram of the SNPs on the array (to identify possible sample
 replications or appraise relatedness), scatterplot of first two PCs,
 boxplots of batch and other characteristics (if indicated) vs the top
 principle components, csv file of suggested probes to remove based on the
 percent of samples that failed for that probe (see function details), csv
 file of suggested samples to potentially remove based on evidence of sex
 mix-ups, a high proportion of failed probes for that sample, or
 unintentional replicates based on clustering of SNP probes (see dendogram of
 the SNPs on the array). Also returns a list that includes:
 \item{SamplestoRemove}{A character vector of Basenames to exclude based on
 evidence of sex mix-ups, or a high proportion of failed probes for that
 sample. This vector does not the suspected unintentional replicates; we
 suggest users check the SNP dendogram to evaluate this themselves.}
 \item{ProbestoRemove}{A character vector of CpGIDs to exclude based on the
 exploratory data analysis} \item{DetectionPval}{A matrix used to mask
 methylation values with a poor detection p-value; NA if poor detection
 p-value and 1 otherwise}
}
\description{
Exploratory data analysis to identify possible sample swaps and
 technical outliers
}
\details{
The sesame package is used to calculate the detection p-values.
 Samples to exclude based on whether they look like a sex mix-up, and poor
 detection p-values (detection p-value>0.05) for more than 5 percent of
 probes.
}
\examples{
\dontrun{
EDAtrying<-ExploratoryDataAnalysis(RGset=exampledat,
                                  globalvarexplore=c("Sex","BWT"),
                                  destinationfolder="H:/UCLA/PACE/Birthweight-placenta",
                                  savelog=TRUE,
                                  cohort="HEBC",analysisdate="20200710")
}
}
